refactor: mapping results processing
